feat(examples): slide puzzle example working with Live S
